finished basic list/get/inspect for volumes
