+++ DockerSdk.Tests/VolumeAccessTests.cs
@@ -3,29 +3,12 @@ using System.Threading.Tasks;
 using DockerSdk.Volumes;
 using FluentAssertions;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace DockerSdk.Tests
 {
     [Collection("Common")]
     public class VolumeAccessTests
     {
-        public VolumeAccessTests(ITestOutputHelper toh)
-        {
-            this.toh = toh;
-        }
-
-        private readonly ITestOutputHelper toh;
-
-        [Fact]
-        public async Task GetDetailsAsync_NoSuchVolume_ThrowsVolumeNotFoundException()
-        {
-            using var client = await DockerClient.StartAsync();
-
-            await Assert.ThrowsAnyAsync<VolumeNotFoundException>(
-                () => client.Volumes.GetDetailsAsync("ddnt-no-such-volume-366ad733152b70e53ddd7fd59defe9fa2e055ed2090f5f3a8839b2797388d0b4"));
-        }
-
         [Fact]
         public async Task GetDetailsAsync_LocalVolumeExists_GetsExpectedDetails()
         {
@@ -41,6 +24,15 @@ namespace DockerSdk.Tests
             volume.Mountpoint.Should().NotBeNullOrWhiteSpace();
             volume.Name.ToString().Should().Be("ddnt-general");
             volume.Scope.Should().Be(VolumeScope.Local);
+        }
+
+        [Fact]
+        public async Task GetDetailsAsync_NoSuchVolume_ThrowsVolumeNotFoundException()
+        {
+            using var client = await DockerClient.StartAsync();
+
+            await Assert.ThrowsAnyAsync<VolumeNotFoundException>(
+                () => client.Volumes.GetDetailsAsync("ddnt-no-such-volume-366ad733152b70e53ddd7fd59defe9fa2e055ed2090f5f3a8839b2797388d0b4"));
         }
 
         [Fact]
