+++ DockerSdk/Volumes/IVolume.cs
@@ -1,32 +1,30 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 
 namespace DockerSdk.Volumes
 {
+    /// <summary>
+    /// Represents a Docker volume, which is a persistent filesystem mount used by containers.
+    /// </summary>
+    /// <seealso cref="IVolumeInfo"/>
     public interface IVolume
     {
+        /// <summary>
+        /// Gets the volume's name.
+        /// </summary>
         VolumeName Name { get; }
 
-        Task<IVolumeInfo> GetDetailsAsync(CancellationToken ct = default);
-    }
-
-    public interface IVolumeInfo : IVolume
-    {
-        DateTimeOffset CreationTime { get; }
-
-        string Driver { get; }
-
         /// <summary>
-        /// Gets the labels that have been applied to the volume.
+        /// Gets detailed information about the volume.
         /// </summary>
-        IReadOnlyDictionary<string, string> Labels { get; }
+        /// <param name="ct">A <see cref="CancellationToken"/> used to cancel the operation.</param>
+        /// <returns>A <see cref="Task"/> that completes when the result is available.</returns>
+        /// <exception cref="VolumeNotFoundException">The volume no longer exists.</exception>
+        /// <exception cref="System.Net.Http.HttpRequestException">
+        /// The request failed due to an underlying issue such as network connectivity, DNS failure, server certificate
+        /// validation, or timeout.
+        /// </exception>
+        Task<IVolumeInfo> GetDetailsAsync(CancellationToken ct = default);
 
-        string Mountpoint { get; }
-
-        VolumeScope Scope { get; }
-
-        // TODO: Options
     }
 }
